AI added
AI add been added with 87% accuracy.
The Ai is serialized inside the folder AISerialized

When we open for the first time the details of an movie, the AI will
analyse all comments and the feedback will be save along with the
comment.

The field ai_feedback can't be 0 or 1. 0 means that the comment are
evaluated as negative and 1 positive. The field Ai_probability_feedback
is the probability of the value "ai_feedback" being true.